+++ server/settings.py
@@ -1,6 +1,4 @@
-global ELASTIC_PASSWORD, CERT_FINGERPRINT, MAX_SIZE
-ELASTIC_PASSWORD = "RrfPwkaofSx_1m6rnmWd"
-CERT_FINGERPRINT = "3ad62a2603d0fe2ecb51038ae775f0bc015e495984ce044ffe54f1722355a421"
-
-kibana = 'eyJ2ZXIiOiI4LjEyLjEiLCJhZHIiOlsiMTcyLjE4LjAuMjo5MjAwIl0sImZnciI6ImUwMWJjZGFhNDU1Y2JhYjUzYmQwODc3NmFhOTlhMzI2M2ZkOGZhNmM4YjE4ZDkzMzk3MmU5MzY0MDg4NjlkMDkiLCJrZXkiOiJ6N0VadlkwQk5neUVfUTd1OXZmNDpZN2o3M2duQ1N1Q3NjRXd0V3JMbjR3In0='
-MAX_SIZE = 15
+global ELASTIC_PASSWORD, CERT_FINGERPRINT, API_KEY
+ELASTIC_PASSWORD = "06Tv9FUtTxQ43sdNuSdU"
+CERT_FINGERPRINT = "e01bcdaa455cbab53bd08776aa99a3263fd8fa6c8b18d933972e936408869d09"
+API_KEY = "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"
